requirements.txt
Flask==2.1.1
discord.py==2.0.0
colorama==0.4.4